refactor: Rename `lib/plugins/aws/common/lib/cleanupTempDir.js`
