+++ lib/plugins/aws/common/index.js
@@ -7,7 +7,7 @@
  */
 
 const validate = require('../lib/validate');
-const cleanupTempDir = require('./lib/cleanupTempDir');
+const cleanupTempDir = require('./lib/cleanup-temp-dir');
 const artifacts = require('./lib/artifacts');
 
 class AwsCommon {
