+++ lib/plugins/aws/common/lib/cleanup-temp-dir.js
