+++ test/unit/lib/plugins/aws/common/lib/cleanup-temp-dir.test.js
